requirements/base.txt
django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-autoslug==1.9.9
django-countries==7.5.1
django-taggit==5.0.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.31
drf-yasg==1.21.7
pillow==10.2.0
argon2-cffi==23.1.0
pytz==2024.1
redis==5.0.2
celery==5.3.6
flower==2.0.1
django-celery-email==3.0.0
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
